refactor: split config.rs into submodules
Extract the 3585-line config.rs into a config/ directory with focused
submodules:

- error.rs: ConfigError enum
- types.rs: Config, Resolved, CliFlags, LoadedConfig structs
- loading.rs: Config::load_file/parse/merge, Resolved impls
- path.rs: config_path, expand_tilde, resolve_config_path
- validation.rs: validate_config, diagnostics, edit distance
- registry.rs: CONFIG_KEYS, lookup_key, ConfigValueType
- display.rs: explain_key, explain_all, display_config
- editing.rs: set/append/remove/unset/get value operations
- repo.rs: RepoKeyTarget, repo config set support
- mod.rs: re-exports only (public API unchanged)

Pure structural refactor — no logic changes. Tests distributed to the
module that defines the tested function.
